APIS-4714 - Factored out acceptance tests to own top level dir
