+++ acceptance/uk/gov/hmrc/apidocumentation/DescriptiveMocks.scala
@@ -14,10 +14,7 @@
  * limitations under the License.
  */
 
-package utils.uk.gov.hmrc.apidocumentation.mocks
-
-import acceptance.uk.gov.hmrc.apidocumentation.specs.ComponentTestsSpec
-import org.scalatest.prop.{TableDrivenPropertyChecks, TableFor1}
+package uk.gov.hmrc.apidocumentation
 
 trait DescriptiveMocks extends Stubs {
 
@@ -54,10 +51,3 @@ trait DescriptiveMocks extends Stubs {
   }
 }
 
-trait TableDrivenPropertyMocks extends TableDrivenPropertyChecks { cs: ComponentTestsSpec =>
-  def helloWorldVersionsIsDeployed(versionTable:TableFor1[String]=Table("Versions", "1.0", "1.2")) {
-    forAll(versionTable) { version =>
-      And.helloWorldIsDeployed("api-example-microservice", version)
-    }
-  }
-}
