+++ acceptance/uk/gov/hmrc/apidocumentation/pages/Section.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package acceptance.uk.gov.hmrc.apidocumentation.pages
+package uk.gov.hmrc.apidocumentation.pages
 
-import acceptance.uk.gov.hmrc.apidocumentation.WebPage
+import uk.gov.hmrc.apidocumentation.WebPage
 import org.openqa.selenium.{By, WebElement}
 
 import scala.util.Try
